+++ program-runtime/src/instruction_processor.rs
@@ -1,15 +1,17 @@
-use crate::native_loader::NativeLoader;
+use crate::{
+    ic_msg,
+    invoke_context::{InvokeContext, ProcessInstructionWithContext},
+    native_loader::NativeLoader,
+};
 use serde::{Deserialize, Serialize};
 use solana_sdk::{
     account::{AccountSharedData, ReadableAccount, WritableAccount},
     account_utils::StateMut,
     bpf_loader_upgradeable::{self, UpgradeableLoaderState},
     feature_set::{demote_program_write_locks, do_support_realloc, remove_native_loader},
-    ic_msg,
     instruction::{Instruction, InstructionError},
     keyed_account::keyed_account_at_index,
     message::Message,
-    process_instruction::{Executor, InvokeContext, ProcessInstructionWithContext},
     pubkey::Pubkey,
     rent::Rent,
     system_instruction::MAX_PERMITTED_DATA_LENGTH,
@@ -18,9 +20,22 @@ use solana_sdk::{
 use std::{
     cell::{Ref, RefCell, RefMut},
     collections::HashMap,
+    fmt::Debug,
     rc::Rc,
     sync::Arc,
 };
+
+/// Program executor
+pub trait Executor: Debug + Send + Sync {
+    /// Execute the program
+    fn execute(
+        &self,
+        first_instruction_account: usize,
+        instruction_data: &[u8],
+        invoke_context: &mut dyn InvokeContext,
+        use_jit: bool,
+    ) -> Result<(), InstructionError>;
+}
 
 #[derive(Default)]
 pub struct Executors {
