+++ programs/bpf/rust/sysvar/Cargo.toml
@@ -13,6 +13,7 @@ edition = "2018"
 solana-program = { path = "../../../../sdk/program", version = "=1.9.0" }
 
 [dev-dependencies]
+solana-program-runtime = { path = "../../../../program-runtime", version = "=1.9.0" }
 solana-program-test = { path = "../../../../program-test", version = "=1.9.0" }
 solana-sdk = { path = "../../../../sdk", version = "=1.9.0" }
