+++ programs/compute-budget/Cargo.toml
@@ -10,6 +10,7 @@ license = "Apache-2.0"
 edition = "2018"
 
 [dependencies]
+solana-program-runtime = { path = "../../program-runtime", version = "=1.9.0" }
 solana-sdk = { path = "../../sdk", version = "=1.9.0" }
 
 [lib]
