+++ programs/config/Cargo.toml
@@ -14,11 +14,11 @@ bincode = "1.3.3"
 chrono = { version = "0.4.11", features = ["serde"] }
 serde = "1.0.130"
 serde_derive = "1.0.103"
+solana-program-runtime = { path = "../../program-runtime", version = "=1.9.0" }
 solana-sdk = { path = "../../sdk", version = "=1.9.0" }
 
 [dev-dependencies]
 solana-logger = { path = "../../logger", version = "=1.9.0" }
-solana-program-runtime = { path = "../../program-runtime", version = "=1.9.0" }
 
 [lib]
 crate-type = ["lib"]
