+++ programs/stake/Cargo.toml
@@ -19,8 +19,8 @@ serde_derive = "1.0.103"
 solana-frozen-abi = { path = "../../frozen-abi", version = "=1.9.0" }
 solana-frozen-abi-macro = { path = "../../frozen-abi/macro", version = "=1.9.0" }
 solana-metrics = { path = "../../metrics", version = "=1.9.0" }
-solana-sdk = { path = "../../sdk", version = "=1.9.0" }
 solana-program-runtime = { path = "../../program-runtime", version = "=1.9.0" }
+solana-sdk = { path = "../../sdk", version = "=1.9.0" }
 solana-vote-program = { path = "../vote", version = "=1.9.0" }
 solana-config-program = { path = "../config", version = "=1.9.0" }
 thiserror = "1.0"
