+++ programs/vote/Cargo.toml
@@ -20,11 +20,9 @@ solana-frozen-abi = { path = "../../frozen-abi", version = "=1.9.0" }
 solana-frozen-abi-macro = { path = "../../frozen-abi/macro", version = "=1.9.0" }
 solana-logger = { path = "../../logger", version = "=1.9.0" }
 solana-metrics = { path = "../../metrics", version = "=1.9.0" }
+solana-program-runtime = { path = "../../program-runtime", version = "=1.9.0" }
 solana-sdk = { path = "../../sdk", version = "=1.9.0" }
 thiserror = "1.0"
-
-[dev-dependencies]
-solana-program-runtime = { path = "../../program-runtime", version = "=1.9.0" }
 
 [build-dependencies]
 rustc_version = "0.4"
