+++ rbpf-cli/src/main.rs
@@ -5,16 +5,16 @@ use solana_bpf_loader_program::{
     create_vm, serialization::serialize_parameters, syscalls::register_syscalls, BpfError,
     ThisInstructionMeter,
 };
-use solana_program_runtime::invoke_context::{prepare_mock_invoke_context, ThisInvokeContext};
+use solana_program_runtime::invoke_context::{
+    prepare_mock_invoke_context, InvokeContext, ThisInvokeContext,
+};
 use solana_rbpf::{
     assembler::assemble,
     static_analysis::Analysis,
     verifier::check,
     vm::{Config, DynamicAnalysis, Executable},
 };
-use solana_sdk::{
-    account::AccountSharedData, bpf_loader, process_instruction::InvokeContext, pubkey::Pubkey,
-};
+use solana_sdk::{account::AccountSharedData, bpf_loader, pubkey::Pubkey};
 use std::{fs::File, io::Read, io::Seek, io::SeekFrom, path::Path};
 use time::Instant;
