+++ sdk/src/lib.rs
@@ -35,7 +35,6 @@ pub mod nonce_account;
 pub mod packet;
 pub mod poh_config;
 pub mod precompiles;
-pub mod process_instruction;
 pub mod program_utils;
 pub mod pubkey;
 pub mod recent_blockhashes_account;
